+++ R/bq-request.R
@@ -31,7 +31,7 @@ bq_ua <- function() {
     "bigrquery/", utils::packageVersion("bigrquery"), " ",
     "(GPN:RStudio; )", " ",
     "gargle/", utils::packageVersion("gargle"), " ",
-    "httr/", utils::packageVersion("httr")
+    "httr2/", utils::packageVersion("httr2")
   )
 }
 
@@ -40,33 +40,29 @@ bq_body <- function(body, ...) {
   utils::modifyList(body, user)
 }
 
-
-#' @importFrom httr GET config
+#' @importFrom httr2 req_perform resp_body_json resp_body_raw
 bq_get <- function(url, ..., query = NULL, raw = FALSE, token = bq_token()) {
-  req <- GET(
-    paste0(base_url, url),
-    token,
-    httr::user_agent(bq_ua()),
-    ...,
-    query = prepare_bq_query(query)
-  )
-  process_request(req, raw = raw)
+  req <- bq_request(url, "GET", token, query = query)
+  resp <- req_perform(req)
+  if (!raw) {
+    resp_body_json(resp)
+  } else {
+    resp_body_raw(resp)
+  }
 }
 
-bq_exists <- function(url, ..., query = NULL, token = bq_token()) {
-  req <- GET(
-    paste0(base_url, url),
-    token,
-    httr::user_agent(bq_ua()),
-    ...,
-    query = prepare_bq_query(query)
-  )
-  status_code(req) >= 200 && status_code(req) < 300
+bq_exists <- function(url, query = NULL, token = bq_token()) {
+  req <- bq_request(url, "GET", token, query = query)
+  resp <- req_perform(req)
+  # A 404 is not an error here.
+  req <- req_error(req, is_error = function(resp) {
+    resp_status(resp) != 404 && resp_status(resp) >= 400
+  })
+  resp <- req_perform(req)
+  resp_status(resp) != 404
 }
 
-
-#' @importFrom httr GET config
-bq_get_paginated <- function(url, ..., query = NULL, token = bq_token(),
+bq_get_paginated <- function(url, query = NULL, token = bq_token(),
                              page_size = 50, max_pages = Inf, warn = TRUE) {
 
   check_number_whole(max_pages, min = 1, allow_infinite = TRUE)
@@ -78,14 +74,14 @@ bq_get_paginated <- function(url, ..., query = NULL, token = bq_token(),
   query <- utils::modifyList(list(maxResults = page_size), query %||% list())
   pages <- list()
 
-  page <- bq_get(url, ..., query = query, token = token)
+  page <- bq_get(url, query = query, token = token)
   i <- 1
   pages[[i]] <- page
   page_token <- page$nextPageToken
 
   while (!is.null(page_token) && i < max_pages) {
     query$pageToken <- page_token
-    page <- bq_get(url, ..., query = query, token = token)
+    page <- bq_get(url, query = query, token = token)
 
     i <- i + 1
     pages[[i]] <- page
@@ -103,137 +99,81 @@ bq_get_paginated <- function(url, ..., query = NULL, token = bq_token(),
   pages
 }
 
-
-#' @importFrom httr DELETE config
-bq_delete <- function(url, ..., query = NULL, token = bq_token()) {
-  req <- DELETE(
-    paste0(base_url, url),
-    token,
-    httr::user_agent(bq_ua()),
-    ...,
-    query = prepare_bq_query(query)
-  )
-  process_request(req)
+#' @importFrom httr2 req_perform
+bq_delete <- function(url, query = NULL, token = bq_token()) {
+  req <- bq_request(url, "DELETE", token, query = query)
+  req_perform(req)
+  invisible(NULL)
 }
 
-#' @importFrom httr POST add_headers config
-bq_post <- function(url, body, ..., query = NULL, token = bq_token()) {
-  json <- jsonlite::toJSON(body, pretty = TRUE, auto_unbox = TRUE)
-
-  req <- POST(
-    paste0(base_url, url),
-    body = json,
-    httr::user_agent(bq_ua()),
-    token,
-    add_headers("Content-Type" = "application/json"),
-    ...,
-    query = prepare_bq_query(query)
-  )
-  invisible(process_request(req))
+#' @importFrom httr2 req_body_json req_perform resp_body_json
+bq_post <- function(url, body, query = NULL, token = bq_token()) {
+  req <- bq_request(url, "POST", token, query = query)
+  req <- req_body_json(req, body)
+  resp <- req_perform(req)
+  invisible(resp_body_json(resp))
 }
 
-#' @importFrom httr PATCH add_headers config
-bq_patch <- function(url, body, ..., query = NULL, token = bq_token()) {
-  json <- jsonlite::toJSON(body, pretty = TRUE, auto_unbox = TRUE)
-  req <- PATCH(
-    paste0(base_url, url),
-    body = json,
-    httr::user_agent(bq_ua()),
-    token,
-    add_headers("Content-Type" = "application/json"),
-    ...,
-    query = prepare_bq_query(query)
-  )
-  process_request(req)
+#' @importFrom httr2 req_body_json req_perform resp_body_json
+bq_patch <- function(url, body, query = NULL, token = bq_token()) {
+  req <- bq_request(url, "PATCH", token, query = query)
+  req <- req_body_json(req, body)
+  resp <- req_perform(req)
+  resp_body_json(resp)
 }
 
-#' @importFrom httr POST PUT add_headers headers config status_code
+#' @importFrom httr2 req_body_json req_body_raw req_perform resp_body_json
 # https://cloud.google.com/bigquery/docs/reference/api-uploads
 bq_upload <- function(url, metadata, media, query = list(), token = bq_token()) {
 
   query <-  utils::modifyList(list(fields = "jobReference",uploadType = "resumable"), query)
-  config <- add_headers("Content-Type" = metadata[["type"]])
 
-  req <- POST(
-    paste0(upload_url, url),
-    body = metadata[["content"]],
-    httr::user_agent(bq_ua()),
-    token,
-    config,
-    query = query
-  )
+  req <- bq_request(url, "POST", token, query = query, base_url = upload_url)
+  req <- req_body_json(req, metadata[["content"]])
+  resp <- req_perform(req)
 
-  if (status_code(req) == 200) {
-
-    config <- add_headers("Content-Type" = media[["type"]])
-
-    req <- PUT(
-      headers(req)$location,
-      body = media[["content"]],
-      httr::user_agent(bq_ua()),
-      token,
-      config
-    )
-
-  }
-
-  process_request(req)
+  # Note: We only get here if the request above is successful.
+  session_uri <- httr2::resp_header(resp, "Location")
+  req <- bq_request(session_uri, "PUT", token, query = query)
+  req <- req_body_raw(req, media[["content"]], type = media[["type"]])
+  resp <- req_perform(req)
+  resp_body_json(resp)
 }
 
-
-#' @importFrom httr http_status content parse_media status_code
-process_request <- function(req, raw = FALSE, call = caller_env()) {
-  status <- status_code(req)
-  # No content -> success
-  if (status == 204) return(TRUE)
-
-  type <- req$headers$`Content-type`
-  content <- content(req, "raw")
-
-  bq_check_response(
-    status = status,
-    type = type,
-    content = content,
-    call = call
-  )
-
-  if (raw) {
-    content
-  } else {
-    jsonlite::fromJSON(rawToChar(content), simplifyVector = FALSE)
+#' @importFrom httr2 request req_user_agent req_url_path_append req_method req_auth_bearer_token req_url_query req_error req_perform
+bq_request <- function(url,
+                       method,
+                       token,
+                       query = NULL,
+                       base_url = base_url,
+                       call = caller_env()) {
+  req <- request(base_url)
+  req <- req_user_agent(req, bq_ua())
+  req <- req_url_path_append(req, url)
+  req <- req_method(req, method)
+  req <- req_auth_bearer_token(req, token)
+  if (!is.null(query)) {
+    req <- req_url_query(req, !!!prepare_bq_query(query))
   }
+  req_error(req, body = bq_error_body)
 }
 
-bq_check_response <- function(status, type, content, call = caller_env()) {
-  if (status >= 200 && status < 300) {
-    return()
+#' @importFrom httr2 resp_content_type resp_status resp_body_json resp_body_string
+bq_error_body <- function(resp) {
+  # Generic error message for non-JSON responses.
+  if (resp_content_type(resp) != "application/json") {
+    message <- paste0(
+      "HTTP error [",
+      resp_status(resp),
+      "]\n",
+      resp_body_string(resp)
+    )
+    return(message)
   }
 
-  type <- httr::parse_media(type)
-  text <- rawToChar(content)
-
-  if (type$complete == "application/json") {
-    json <- jsonlite::fromJSON(text, simplifyVector = FALSE)
-    gargle_abort(
-      reason = json$error$errors[[1L]]$reason,
-      message = json$error$message,
-      status = status,
-      call = call
-    )
-  } else {
-    message <- paste0("HTTP error [", status, "]\n", text)
-    gargle_abort(
-      reason = NULL,
-      message = message,
-      status = status,
-      call = call
-    )
-  }
-}
-
-gargle_abort <- function(reason, message, status, call = caller_env()) {
-  class <- paste0("bigrquery_http_", status)
-
+  body <- resp_body_json(resp)
+  reason <- body$error$errors[[1L]]$reason
+  message <- body$error$message
   if (!is.null(reason)) {
     advice <- NULL
     if (reason == "responseTooLarge") {
@@ -255,9 +195,6 @@ gargle_abort <- function(reason, message, status, call = caller_env()) {
       paste0(message, " [", reason, "] "),
       i = advice
     )
-    class <- c(paste0("bigrquery_", reason), class)
   }
-
-  cli::cli_abort(message, class = class, call = call)
+  message
 }
-